separate rails and webpack configs
instead of trying to share a .env

- rails configuration is in application.rb as config.webpack
- main webpack configuration is in webpack.common.config.js
