--- .env
@@ -1,1 +1,0 @@
-WEBPACK_DEV_SERVER_PORT = '3030'
