+++ app/helpers/application_helper.rb
@@ -5,11 +5,17 @@ module ApplicationHelper
     stylesheet_link_tag(*sources)
   end
 
-  # prefix development javascript urls with /dev-assets path
+  # use full url to webpack dev-server for javascripts in development
   def env_javascript_include_tag(*sources)
     if Rails.env.development?
       filename, *rest = sources
-      dev_sources = ["/dev-assets/#{filename}", *rest]
+      dev_server_config = Rails.configuration.web_pack[:dev_server]
+      dev_filename = URI::HTTP.build(
+        host: 'localhost',
+        port: dev_server_config[:port],
+        path: "/#{dev_server_config[:path]}/#{filename}.js"
+      ).to_s
+      dev_sources = [dev_filename, *rest]
       javascript_include_tag(*dev_sources)
     else
       javascript_include_tag(*sources)
