+++ config/application.rb
@@ -29,8 +29,17 @@ module RailsWebpack
     # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
     # config.i18n.default_locale = :de
 
-    # destination for webpack asset builds
-    config.assets.paths << Rails.root.join('dist')
+    # webpack configuration
+    config.web_pack = {
+      dev_server: {
+        path: 'dev-assets',
+        port: '3030',
+      },
+      dist_path: Rails.root.join('dist')
+    }
+
+    # add destination for webpack asset builds
+    config.assets.paths << config.web_pack[:dist_path]
 
     # Do not swallow errors in after_commit/after_rollback callbacks.
     config.active_record.raise_in_transactional_callbacks = true
