+++ frontend/config/webpack.common.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+  DEV_SERVER: {
+    PORT: '3030',
+    PATH: 'dev-assets'
+  },
+  DIST_PATH: './dist',
+  ENTRY: {
+    main: './frontend/app/index.js'
+  },
+  FRONTEND_PATH: './frontend/app',
+};
