+++ frontend/config/webpack.prod.config.js
@@ -1,7 +1,7 @@
 var webpack = require('webpack');
 var path = require('path');
+var config = require('./webpack.common.config');
 var ExtractTextPlugin = require('extract-text-webpack-plugin');
-var config = require('./webpack.shared');
 
 module.exports = {
